Fixed coverage checker
Dealing with new source files not part of the reference data
caused an exception in the checker. This is fixed treating new
source files as 100% covered which are then compared with their
real coverage value. If the new source can't be 100% covered
an update of the reference data is needed via 'make coverage'