Fixed config parser default section handling
The DEFAULT section in an INI file is handled special by
python's config parser implementation. Therefore this patch
refactors the way we use the config file sections. Any
section of the config file belongs to an account, the DEFAULT
section allows to select one of those accounts as the default
account with the default_account keyword. In addition the
subcommand to handle the default account was added. This
fixes Issue #51