Change configuration file structure
Instead of one account section containing information about the Azure
account as well as information about the storage/container account it
is required to handle these information in separate sections. The reason
for this change is that storage accounts can exist in different regions.
Therefore the storage accounts should be placed in a region section
This refactoring also impacts commands with --region and/or --container
options because they are now handled as global options.

The new configuration format is as follows:

  [DEFAULT]
  default_account = account:name
  default_region  = region:name

  [account:some-account]
  publishsettings = filename

  [region:some-region]
  default_storage_account   = name
  default_storage_container = name
  storage_accounts   = comma-list
  storage_containers = comma-list

The account setup command allows to perform the following actions

  * add account and region sections
  * remove sections
  * set default account
  * set default region

There is still room for improvement on the account setup command
However that should be part of an additional implementation. This
fixes Issue #57