Delete config file options
storage_accounts and storage_containers are gone from the config
file. They were only used for pre validation of specified storage
account and container names. This validation is meaningless if
the data provided in the config file is not valid compared to the
service. In addition the combination of storage account and
container was not validated at all. Thus these data was removed
from the config file. If an invalid storage account or container
or a combination of both is used azurectl will raise an exception
with the data we get back from the service. This fixes Issue #61