Added check for cloudservice URL
A requested cloud service will be registered as public address in Azure
If this address is already in use the service comes back with a message
that is not clear to the user how to fix the situation. This patch adds
a DNS check prior to creating a new cloud service and provides a better
error message to the user. This fixes Issue #67