Create a uniq disk identifier on vm create
When instantiating a new VM, we create a disk file with the
name '_instance', in the selected storage container. Since
storage files have unique names an additional information
is required to allow creation of multiple vm's from the same image
in a selected region and container. The naming schema is as
follows

    <cloud-service-name>_instance_<instance-name>_image_<image-name>

This fixes #99