Allow setup of custom Azure SDK in virtualenv
If the pip version of the Azure SDK and the Azure Storage SDK
is not suitable for the development tasks we have to manually
install the code from git into the virtualenv in order to allow
azurectl to make use of this SDK version. This commit adds the
helper script .virtualenv.install-azure-sdk which pulls the
SUSE SDK fork to the local development tree.