Not all operations requires a region setup
azurectl operations like 'storage account list' or 'compute vm types'
as well as 'compute vm regions' do not require a region setup in the
configuration file. In addition we allow to setup the account with
the region setup to be optional but the code always reads in a default
region setup. Therefore this commit changes the behaviour to read
the region setup only when required and allows for operations without
a region setup. This is related to Issue #124