Allow a region argument to bypass the config file
There are specific cases where the region argument is valuable
in a function, but not essential to the configuration file.

An example is `storage account create`; it's necessary to define
the region where the storage account will be placed, but a _specific
storage account_ or a _specific container_ is not required (which is
what the config provides); only the name of the region.

In the event the region name is invalid, it will be reflected in an
error raised by the upstream API.