Added --wait option for async commands
Commands which comes back with a request id can be started
with the --wait option. This causes the command to block the
caller until the request status changes to a success state.
There are commands like replication which implements the
wait differently because the condition of success for the
request is not enough. What exactly the command is waiting
for needs to end up in the documentation