Update bash completion
On make build the bash completion is automatically updated
to the latest version. Thus it's not required to keep an
up to date version of the file in the git repo. However,
it also does not hurt and allows to make use of the latest
completion from within a development environment