Fix methods if no endpoints.
Azure is returning a None type instead of an empty list if no
endpoints exist. Check for None in show and list method.

In create method create a new instance of endpoints list if
it is None for appending the new endpoint.

In delete return an error if the endpoint is not foud in the list
or if there are no endpoints.

In show endpoint method return from for loop.

If endpoint is found return from for loop. Otherwise if loop
finishes raise exception, endpoint not found.