Refactor XZ reading of size bytes
Implement a read buffer for XZ decompression. It's required that we
can read an exact number of uncompressed bytes from the XZ stream.
As the max_length parameter was added to the lzma.decompress()
method with python 3.5 for the first time and we have to stay
compatible with 3.4 a self written solution to this problem was
required