Refactor handling of request IDs
Added unify_id method for request id validation
The request id value from the Azure rest api is not consistently
of the same type which makes it hard to handle the id if it could
be a string or a number. This commit introduce a static method
called unify_id which allows us to properly format the id into
a common value so that azurectl can consume the information in
the same way all over the place. So far unify_id just python
formats the information into a string.

For the handling of the cloud service request id an explicit
conversion into a number is performed in addition to the
refactoring code